@@ -1,4 +1,4 @@
-package co.devskills.mobile_boilerplate_flutter
+package io.alvalabs.mobile_boilerplate_flutter
 
 import io.flutter.embedding.android.FlutterActivity
 
